nothing of note

+++ src/main.cpp
@@ -110,8 +110,8 @@ void loop() {
       connect_wait();
     }  
     digitalWrite(LED,ON);
-    int x = (2750-analogRead(VRX))/12;
-    int y = (2730-analogRead(VRY))/12;
+    int8_t x = (int8_t)((2750-analogRead(VRX))/12);
+    int8_t y = (int8_t)((2730-analogRead(VRY))/12);
 
     x /= (x>0) ? 2 : 1;
     y /= (y>0) ? 2 : 1;
